+++ pkg/distributor/push_test.go
@@ -1359,6 +1359,10 @@ func (o otlpLimitsMock) PromoteOTelResourceAttributes(string) []string {
 	return nil
 }
 
+func (o otlpLimitsMock) OTelKeepIdentifyingResourceAttributes(string) bool {
+	return false
+}
+
 func promToMimirHistogram(h *prompb.Histogram) mimirpb.Histogram {
 	pSpans := make([]mimirpb.BucketSpan, 0, len(h.PositiveSpans))
 	for _, span := range h.PositiveSpans {